backend/requirements.txt
Django
solana
solders
python-dotenv